Counts shit
Code successfully runs on a small simple test file.  Accurately counts
the amount of letters and punctuation.  Spaces are a little more tricky
because I don't count the spaces after punctuation.  The space counter
register basically represents the spaces between words.  That means if
we add punctuation and spaces we have the total number of words.  And
punctuation represent the total number of sentences.